// packages/core/src/types/types.ts
export interface Link<T extends string> {
  sys: {
    type: 'Link';
    linkType: T;
    id: string;
  };
}
// ...
export type DataSourceEntryValueType = Link<'Entry' | 'Asset'>;

export type CompositionVariableValueType = string | boolean | number | Record<any, any> | undefined;
type CompositionComponentPropType =
  | 'BoundValue'
  | 'UnboundValue'
  | 'DesignValue'
  | 'ComponentValue';

export type CompositionComponentPropValue<
  T extends CompositionComponentPropType = CompositionComponentPropType
> = T extends 'DesignValue'
  ? // The keys in valuesByBreakpoint are the breakpoint ids
    { type: T; valuesByBreakpoint: Record<string, CompositionVariableValueType> }
  : T extends 'BoundValue'
  ? { type: T; path: string }
  : { type: T; key: string };

// TODO: add conditional typing magic to reduce the number of optionals
export type CompositionComponentNode = {
  type: 'block' | 'root' | 'editorRoot' | 'designComponent' | 'designComponentBlock';
  data: {
    id: string;
    blockId?: string; // will be undefined in case string node or if root component
    props: Record<string, CompositionComponentPropValue<CompositionComponentPropType>>;
    dataSource: CompositionDataSource;
    unboundValues: CompositionUnboundValues;
    breakpoints: Breakpoint[];
  };
  children: CompositionComponentNode[];
  parentId?: string;
};
// ...
export type CompositionNode = {
  definitionId: string;
  children: Array<CompositionNode>;
  variables: Record<string, CompositionComponentPropValue>;
};

export type CompositionDataSource = Record<string, DataSourceEntryValueType>;
export type CompositionUnboundValues = Record<string, { value: CompositionVariableValueType }>;

export type Breakpoint = {
  id: string;
  query: string;
  displayName: string;
  previewSize: string;
};

export type SchemaVersions = '2023-09-28' | '2023-06-27' | '2023-07-26' | '2023-08-23';

export type ExperienceComponentSettings = {
  variableDefinitions: Record<
    string,
    Omit<ComponentDefinitionVariableBase<ComponentDefinitionVariableType>, 'defaultValue'> & {
      defaultValue: CompositionComponentPropValue<'BoundValue' | 'UnboundValue'>;
    }
  >;
};

export type Composition = {
  title: string;
  slug: string;
  componentTree: {
    breakpoints: Array<Breakpoint>;
    children: Array<CompositionNode>;
    schemaVersion: SchemaVersions;
  };
  dataSource: CompositionDataSource;
  unboundValues: CompositionUnboundValues;
  usedComponents?: Array<Link<'Entry'> | ExperienceEntry>;
  componentSettings?: ExperienceComponentSettings;
};

export type ExperienceEntry = {
  sys: Entry['sys'];
  fields: Composition;
  metadata: Entry['metadata'];
};